move time type extensions to `janus_core`
As part of #990, we want to a field `interval: (chrono::DateTime<Utc>,
chrono::Duration)` to `janus_collector::Collection`. So we will need to
be able to convert from `janus_messages::Time` to `chrono::DateTime<Utc>`
and `janus_messages::Duration` to `chrono::Duration`. This commit moves
the `janus_aggregator::TimeExt`, `janus_aggregator::DurationExt` and
`janus_aggregator::IntervalExt` traits and impls into `janus_core` so
they can be used from `janus_collector`.

While we're at it, we disable the `oldtime` feature on `chrono` as
suggested by the docs ([1]), and hoist the `chrono` dep up into the
workspace `Cargo.toml`. Disabling default `chrono` features doesn't
actually achieve much, because those features get turned back on by
other Janus dependencies that use `chrono`, but at least it's not our
fault that `wasm-bindgen` and `winapi` get pulled in.

[1]: https://docs.rs/chrono/latest/chrono/index.html#duration
