+++ Cargo.toml
@@ -21,6 +21,10 @@ rust-version = "1.65.0"
 version = "0.3.0"
 
 [workspace.dependencies]
+# Disable default features to disable compatibility with the old `time` crate, and we also don't
+# (yet) need other default features.
+# https://docs.rs/chrono/latest/chrono/#duration
+chrono = { version = "0.4", default-features = false }
 janus_aggregator = { version = "0.3", path = "aggregator" }
 janus_build_script_utils = { version = "0.3", path = "build_script_utils" }
 janus_client = { version = "0.3", path = "client" }
