+++ aggregator/Cargo.toml
@@ -41,7 +41,7 @@ atty = "0.2"
 backoff = { version = "0.4.0", features = ["tokio"] }
 base64 = "0.21.0"
 bytes = "1.4.0"
-chrono = "0.4"
+chrono.workspace = true
 clap = { version = "4.1.8", features = ["derive", "env"] }
 console-subscriber = { version = "0.1.8", optional = true }
 deadpool = { version = "0.9.5", features = ["rt_tokio_1"] }
