+++ aggregator/src/aggregator/collection_job_driver.rs
@@ -713,7 +713,6 @@ mod tests {
             test_util::ephemeral_datastore,
             Datastore,
         },
-        messages::TimeExt,
         task::{test_util::TaskBuilder, QueryType, Task},
     };
     use assert_matches::assert_matches;
@@ -725,7 +724,7 @@ mod tests {
             install_test_trace_subscriber,
             runtime::TestRuntimeManager,
         },
-        time::{Clock, MockClock, TimeExt as CoreTimeExt},
+        time::{Clock, MockClock, TimeExt},
         Runtime,
     };
     use janus_messages::{
