+++ aggregator/src/aggregator/query_type.rs
@@ -5,7 +5,6 @@ use crate::{
         models::{AggregateShareJob, BatchAggregation, LeaderStoredReport},
         Transaction,
     },
-    messages::TimeExt as _,
     task::Task,
 };
 use async_trait::async_trait;
