+++ core/Cargo.toml
@@ -32,7 +32,7 @@ test-util = [
 [dependencies]
 backoff = { version = "0.4.0", features = ["tokio"] }
 base64 = "0.21.0"
-chrono = "0.4"
+chrono.workspace = true
 derivative = "2.2.0"
 futures = "0.3.26"
 hex = "0.4"
